doc: Improve config
- improve wheelchair config
- extract transfer config
- fix TODOs

# Conflicts:
#	docs/RouteRequest.md
